Updated tsconfig/tslint to support unused vars, and added commandbar styles
