--- packages/experiments/src/components/CommandBar/CommandBar.scss
@@ -1,19 +1,0 @@
-@import '~office-ui-fabric-react/src/common/common';
-
-.root {
-  display: flex;
-  background-color: $ms-color-neutralLighter;
-  padding: 0 16px;
-}
-
-.primarySet {
-  flex-grow: 1;
-  display: flex;
-  align-items: stretch;
-}
-
-.secondarySet {
-  flex-shrink: 0;
-  display: flex;
-  align-items: stretch;
-}
