+++ packages/experiments/tslint.json
@@ -2,5 +2,25 @@
   "extends": [
     "office-ui-fabric-react-tslint"
   ],
-  "rules": {}
+  "rules": {
+    "jsx-no-string-ref": false,
+    "jsx-key": false,
+    "jsx-boolean-value": false,
+    "jsx-ban-props": false,
+    "max-line-length": [
+      false,
+      140
+    ],
+    "no-any": false,
+    "variable-name": false,
+    "no-internal-module": false,
+    "prefer-const": false,
+    "typedef": [
+      false
+    ],
+    "no-inferrable-types": [
+      false
+    ],
+    "no-unused-variable": false
+  }
 }
